feat: api-handlers-decomposition-001 - Preserve shared API error and JSON response behavior
