+++ pkg/api/handlers.go
@@ -1326,53 +1326,6 @@ func resourceUsagePtr(values []factoryapi.ResourceUsage) *[]factoryapi.ResourceU
 	return &values
 }
 
-func (s *Server) writeJSON(w http.ResponseWriter, status int, v any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	if err := json.NewEncoder(w).Encode(v); err != nil {
-		s.logger.Error("encode response failed", zap.Error(err))
-	}
-}
-
-func (s *Server) writeError(w http.ResponseWriter, status int, message, code string) {
-	s.writeErrorWithTargets(w, status, message, code, nil)
-}
-
-func (s *Server) writeErrorWithTargets(w http.ResponseWriter, status int, message, code string, targets []factoryapi.FactoryValidationTarget) {
-	var targetPtr *[]factoryapi.FactoryValidationTarget
-	if len(targets) > 0 {
-		targetPtr = &targets
-	}
-	s.writeJSON(w, status, factoryapi.ErrorResponse{
-		Message: message,
-		Family:  errorFamilyForStatus(status),
-		Code:    factoryapi.ErrorResponseCode(code),
-		Targets: targetPtr,
-	})
-}
-
-func (s *Server) writeSSEDataJSON(w http.ResponseWriter, v any) error {
-	payload, err := json.Marshal(v)
-	if err != nil {
-		return err
-	}
-	_, err = fmt.Fprintf(w, "data: %s\n\n", payload)
-	return err
-}
-
-func errorFamilyForStatus(status int) factoryapi.ErrorFamily {
-	switch status {
-	case http.StatusBadRequest:
-		return factoryapi.ErrorFamilyBadRequest
-	case http.StatusConflict:
-		return factoryapi.ErrorFamilyConflict
-	case http.StatusNotFound:
-		return factoryapi.ErrorFamilyNotFound
-	default:
-		return factoryapi.ErrorFamilyInternalServerError
-	}
-}
-
 func stringValue(value *string) string {
 	return optional.StringValue(value)
 }
